+++ src/PhpSpreadsheet/Worksheet/Worksheet.php
@@ -307,11 +307,6 @@ class Worksheet
     private ?Color $tabColor = null;
 
     /**
-     * Hash.
-     */
-    private int $hash;
-
-    /**
      * CodeName.
      */
     private ?string $codeName = null;
@@ -323,7 +318,6 @@ class Worksheet
     {
         // Set parent and title
         $this->parent = $parent;
-        $this->hash = spl_object_id($this);
         $this->setTitle($title, false);
         // setTitle can change $pTitle
         $this->setCodeName($this->getTitle());
@@ -378,11 +372,6 @@ class Worksheet
 
         $this->disconnectCells();
         unset($this->rowDimensions, $this->columnDimensions, $this->tableCollection, $this->drawingCollection, $this->chartCollection, $this->autoFilter);
-    }
-
-    public function __wakeup(): void
-    {
-        $this->hash = spl_object_id($this);
     }
 
     /**
@@ -3189,7 +3178,7 @@ class Worksheet
 
         if ($namedRange->getLocalOnly()) {
             $worksheet = $namedRange->getWorksheet();
-            if ($worksheet === null || $this->hash !== $worksheet->getHashInt()) {
+            if ($worksheet === null || $this !== $worksheet) {
                 if ($returnNullIfInvalid) {
                     return null;
                 }
@@ -3333,9 +3322,14 @@ class Worksheet
         return $this;
     }
 
+    /**
+     * @deprecated 5.2.0 Serves no useful purpose. No replacement.
+     *
+     * @codeCoverageIgnore
+     */
     public function getHashInt(): int
     {
-        return $this->hash;
+        return spl_object_id($this);
     }
 
     /**
@@ -3722,7 +3716,6 @@ class Worksheet
                 }
             }
         }
-        $this->hash = spl_object_id($this);
     }
 
     /**